Rename `shortcodesearchresult` package to `identifiersearchresult`
