+++ app/src/main/java/org/simple/clinic/home/HomeScreen.kt
@@ -38,7 +38,7 @@ import org.simple.clinic.router.ScreenResultBus
 import org.simple.clinic.scanid.ScanSimpleIdScreen
 import org.simple.clinic.search.PatientSearchScreenKey
 import org.simple.clinic.settings.SettingsScreenKey
-import org.simple.clinic.shortcodesearchresult.IdentifierSearchResultScreenKey
+import org.simple.clinic.identifiersearchresult.IdentifierSearchResultScreenKey
 import org.simple.clinic.summary.OpenIntention
 import org.simple.clinic.summary.PatientSummaryScreenKey
 import org.simple.clinic.util.UtcClock
