+++ app/src/main/java/org/simple/clinic/identifiersearchresult/IdentifierSearchResultEffect.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.shortcodesearchresult
+package org.simple.clinic.identifiersearchresult
 
 import java.util.UUID
