+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchScreen.kt
@@ -44,7 +44,7 @@ import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.router.ScreenResultBus
 import org.simple.clinic.scanid.ScanSimpleIdScreen
 import org.simple.clinic.scanid.ScanSimpleIdScreenKey
-import org.simple.clinic.shortcodesearchresult.IdentifierSearchResultScreenKey
+import org.simple.clinic.identifiersearchresult.IdentifierSearchResultScreenKey
 import org.simple.clinic.summary.OpenIntention
 import org.simple.clinic.summary.PatientSummaryScreenKey
 import org.simple.clinic.util.RequestPermissions
